all manage device spec stuff completely out of flexsea_device_spec.c
