+++ device_specs/exo_device_spec.c
@@ -1,11 +1,13 @@
 #include "flexsea_device_spec.h"
-#include "flexsea_device_spec_defs.h"
 #include "flexsea_dataformats.h"
 #include "flexsea_user_structs.h"
 #include "flexsea_board.h"
 #include "user-mn-DpEb42.h"
 
-const char* _rigid_m7_fieldlabels[_rigid_m7_numFields] =
+#define _rigid_m7_numFields 36
+#define	_dev_numFields _rigid_m7_numFields
+
+const char* device_spec_labels[_dev_numFields] =
 														{"rigid", 			"id",													// METADATA			2 2
 														"state_time",																// STATE TIME		1 3
 														"accelx", 	"accely", 	"accelz", 	"gyrox", 	"gyroy", 	"gyroz",		// IMU				6 9
@@ -21,7 +23,7 @@ const char* _rigid_m7_fieldlabels[_rigid_m7_numFields] =
 														"cur_stpt",	"step_energy", "walking_state", "gait_state" 					// CONTROLLER		4 36
 };
 
-const uint8_t _rigid_m7_field_formats[_rigid_m7_numFields] =
+const uint8_t device_spec_formats[_dev_numFields] =
 														{FORMAT_8U, 	FORMAT_16U,													// METADATA			2 2
 														FORMAT_32U,																	// STATE TIME		1 3
 														FORMAT_16S, FORMAT_16S, FORMAT_16S, FORMAT_16S, FORMAT_16S, FORMAT_16S ,	// IMU				6 9
@@ -41,7 +43,7 @@ const uint8_t _rigid_m7_field_formats[_rigid_m7_numFields] =
 #define RGD_STRUCT dpRigid
 
 // only defined on boards not on plan
-uint8_t* _rigid_m7_field_pointers[_rigid_m7_numFields] =
+uint8_t* device_spec_variables[_dev_numFields] =
 														{	0,	0,																							// METADATA			2 2
 														PTR2(RGD_STRUCT.ctrl.timestamp),																	// STATE TIME		1 3
 														(uint8_t*)&RGD_STRUCT.mn.accel.x, (uint8_t*)&RGD_STRUCT.mn.accel.y, (uint8_t*)&RGD_STRUCT.mn.accel.z,	// IMU				3 6
@@ -62,15 +64,47 @@ uint8_t* _rigid_m7_field_pointers[_rigid_m7_numFields] =
 };
 
 FlexseaDeviceSpec fx_rigid_m7_spec = {
-		.numFields = _rigid_m7_numFields,
-		.fieldLabels = _rigid_m7_fieldlabels,
-		.fieldTypes = _rigid_m7_field_formats
+		.numFields = _dev_numFields,
+		.fieldLabels = device_spec_labels,
+		.fieldTypes = device_spec_formats
 };
 
 uint32_t *fx_dev_timestamp = &RGD_STRUCT.ctrl.timestamp;
 const FlexseaDeviceSpec *fx_this_device_spec = &fx_rigid_m7_spec;
-const uint8_t ** _dev_data_pointers = (const uint8_t **) _rigid_m7_field_pointers;
-const uint8_t * _dev_field_formats = _rigid_m7_field_formats;
+const uint8_t **_dev_data_pointers = (const uint8_t **) device_spec_variables;
+const uint8_t *_dev_field_formats = device_spec_formats;
 
 #undef RGD_STRUCT
 #undef PTR2
+
+// initialization goes in payload_ptr initialization which is a hack :(
+
+FlexseaDeviceSpec deviceSpecs[NUM_DEVICE_TYPES];
+
+uint32_t fx_active_bitmap[FX_BITMAP_WIDTH_C];
+uint16_t fx_num_fields_active = 0;
+
+const uint8_t *_device_active_field_pointers[_dev_numFields];
+uint8_t _device_active_field_lengths[_dev_numFields];
+
+const uint16_t *read_num_fields_active = &fx_num_fields_active;
+const uint8_t **read_device_active_field_pointers = (const uint8_t **)_device_active_field_pointers;
+const uint8_t *read_device_active_field_lengths =_device_active_field_lengths;
+
+void setActiveFieldsByMap(uint32_t *map)
+{
+	const uint8_t * dev_field_formats = fx_this_device_spec->fieldTypes;
+
+	int i, j=0;
+	for(i=0;i<_dev_numFields; ++i)
+	{
+		if(IS_FIELD_HIGH(i, map))
+		{
+			_device_active_field_pointers[j] = _dev_data_pointers[i];
+			_device_active_field_lengths[j] = FORMAT_SIZE_MAP[dev_field_formats[i]];
+			++j;
+		}
+	}
+
+	fx_num_fields_active = j;
+}
